Changed HomeWork 4. Task 2 and 3 have tests written with pytest, while tasks 3 and 4 were tested with doctest.
Changed solutions for tasks 3,4 and 5 according to your comments.
Changed test for task 2, but instead of using setup&teardown fixtures I've written a static function,
because I can't figure out how to use parametrized fixtures.
I need them parametrized because the fixture must rewrite the file with different text for each test.

And I haven't tested the first task because 'the faulty case' is the example case. The script doesn't seem to work. So I probably think that there's no need in testing not working scripts.